+++ tool/bundle-webpack-module.sh
@@ -1,8 +1,6 @@
 #!/usr/bin/env bash
 
 [[ -f "$MODULE_BUNDLE_SETTING_FILE_PATH" ]] &&	\
-npm install --no-save webpack &&				\
-npm install --no-save webpack-cli &&			\
-npm link webpack &&								\
-npm link webpack-cli &&							\
+npm install --global --silent webpack &&		\
+npm install --global --silent webpack-cli  &&	\
 webpack build --config "$MODULE_BUNDLE_SETTING_FILE_PATH";