+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 #	CHANGELOG
 
-##	2023-11-06/0.1.0
+##	2023-11-15/0.1.0
 
 ###	UNRELEASED
 - undefined
 
 ###	ADDED
-- undefined
+- Improved build procedure.
 
 ###	MODIFIED
 - undefined